Bug 1035442 - Split up collections tests r=amirn
