+++ apps/verticalhome/test/marionette/.jshintrc
@@ -1,6 +1,7 @@
 {
   "extends": "../../../../.jshintrc",
   "predef": [
+    "__dirname",
     "suite",
     "test",
     "setup",
